% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CreateModelFromStr}
\alias{CreateModelFromStr}
\alias{CreateModel}
\title{R wrapper to create GASModel objects.}
\usage{
CreateModelFromStr(model, params = NA)

CreateModel(model, params = NA)
}
\arguments{
\item{model}{String or GASModel object.}

\item{params}{Vector or list of initial or new parameters. If NA and the
\code{model} argument is a string the model's default parameters are used.
If NA and the \code{model} argument is a GASModel object the object's current
parameters are maintained.}
}
\description{
Generates GASModel objects from a string, or wraps \code{GASModel$SetParams}
function if the passed object is of class \code{GASModel}.
}
\section{Functions}{
\itemize{
\item \code{CreateModelFromStr}: Used exclusively with sring argument.
}}

