% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_gas.R
\name{GASModelIndex}
\alias{GASModelIndex}
\alias{GASModel}
\alias{Rcpp_GASModel-class}
\alias{Rcpp_GASModel}
\alias{BetaGenTEGARCH}
\alias{Rcpp_BetaGenTEGARCH-class}
\alias{Rcpp_BetaGenTEGARCH}
\alias{BetaTEGARCH}
\alias{Rcpp_BetaTEGARCH-class}
\alias{Rcpp_BetaTEGARCH}
\alias{DPMP}
\alias{Rcpp_DPMP-class}
\alias{Rcpp_DPMP}
\title{Generalized Autoregressive Score (GAS) models}
\description{
The GAS framework as introduced in introduced in Creal, Lucas & Koopman
(2011), provides a lot of structure, which all models within the framework
share. The implementations of the models in this package reflect
this structure through class ineheritance. The base \code{GASmodel} class
implements a significant share of the functionality and also functions as an
interface. The specific GAS model classes implement the details of the model
specific probability density and other desired specialized functionality.
}
\section{Supported Models}{

\itemize{
 \item{\code{BetaGenTEGARCH}:} {The Beta-Gen-t-EGARCH model introduced by
 Harvey & Lange (2017).}
 \item{\code{BetaTEGARCH}:} {The Beta-t-EGARCH model introduced by Harvey &
 Chakravarty (2008).}
 \item{\code{DPMP}, \code{DPMP1-I}, \code{DPMP1-H}, \code{DPMP1-Inv},
   \code{DPMP2-I}, \code{DPMP2-H}, \code{DPMP2-Inv},
   \code{DPMP3-I}, \code{DPMP3-H}, \code{DPMP3-Inv}:} {
   The Dynamic Pooled Marked Point Process (DPMP) models introduced in Creal,
   Lucas & Koopman (2011). The number represents the number of time-varying
   factors and the specification following the hyphen reflects the scaling
   factor used in the score update; "I" for Identity scaling, "H" for the
   inverse square root of the Fisher Information (FI) matrix, "Inv" for the
   inverse Fisher Information (FI) matrix.}
}
}

\section{Usage}{

\preformatted{
m <- new(GASModel, modelStr)
m <- new(GASModel, modelStr, initParams)
m <- new(GASModel, modelStr, initParams, priorStack)

m <- new(BetaTEGARCH, initParams)
m <- new(BetaTEGARCH, initParams, priorStack)

m$SetParams(params)

m$LogL(y, f1)
m$LogLWPar(params, y, f1)
m$LogPosteriorWPar(params, y, f1)
m$GradLogLWPar(params, y, f1)

m$Filter(y, f1)
}
}

\section{Details}{


\code{new(GASModel, modelStr)},
\code{new(GASModel, modelStr, initParams)},
\code{new(GASModel, modelStr, initParams, priorStack)}.
Create a new GASModel instance of type \code{GASModel}. The
instance has access to the functionality associated with the model type
specified in modelStr. The string modelStr has to be one of the models
included in the supported model list. Optionally
supply a parameter vector \code{initParams} and a prior specification
\code{priorStack} (see \code{\link{PriorStack}})

\code{new(Class, initParams)},
\code{new(Class, initParams, priorStack)}
Create a new GASModel instance of type \code{Class} where \code{Class} is
one of the models included in the supported model list. Optionally
supply a parameter vector \code{initParams} and a prior specification
\code{priorStack}

\code{$SetParams(params)}
Sets the model specific Parameters with the numeric vector \code{params}.

\code{$LogL(y, f1)}
Computes the models log likelihood for data \code{y} and the initial state
for the time-varying parameters set to \code{f1}.

\code{$LogLWPar(y, f1, params)}
Same as \code{$LogL}, but is combined with a call to \code{SetParams} to
update the models parameters with \code{params}.

\code{$LogPosteriorWPar(y, f1, params)}
Computes the models log posterior, which is defined as the sum of log
likelihood plus log prior for GAS models. Like for \code{$LogLWPar} the
model parameters are updated with \code{params}.

\code{$GradLogLWPar(params, y, f1)}
Computes the gradient of the models log likelhood, given the data \code{y},
the initial time-varying paramater state \code{f1}, and model parameters
\code{params}.

\code{$Filter(y, f1)}
Computes the set of time-varying parameters for each observation, given the
data \code{y}, the initial time-varying paramater state \code{f1}. Returns
array like of dimension T by d, where T is the number of observations and d
is the number time-varying parameters of the model.
}

\references{
{
Creal, D., Koopman, S. J., & Lucas, A. (2011b).
 Generalized autoregressive score models with applications [Working paper].

Harvey, A., & Chakravarty, T. (2008).
 Beta-t-(E)GARCH [Cambridge Working Papers in Economics].

Harvey, A., & Lange, R. J. (2017).
 Volatility modeling with a Generalized t distribution.
 Journal of Time Series Analysis, 38(2), 175–190.
}
}
