% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Mode}
\alias{Mode}
\title{Computes the mode.}
\usage{
Mode(x)
}
\arguments{
\item{x}{Vector representing a sample.}
}
\description{
Returns the mode of a sample \code{x} using 0.01 width bins.
}
\examples{
Mode(c(1, 1.8,2, 1.996, 1.991, 2.0005, 2.3, 2.1, 2.5)) # 2

}
