% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.R
\name{PlotACFs}
\alias{PlotACFs}
\title{Autocorrelation function plots of MCMC samples}
\usage{
PlotACFs(samples, mcmcNames, parNames, burn = 1000, lags = 50)
}
\arguments{
\item{samples}{Matrix containing MCMC samples. Each column contains a
seperate chain of a unique combination of a MCMC method specified in
\code{mcmcNames} and parameter specified in \code{parNames}. Columns
must be named according to the convention "\code{mcmcName} \code{parName}"
(e.g. "RWMH A").}

\item{mcmcNames}{Vector or list of strings or epxressions of MCMC method
names. Must be list if expressions are used.}

\item{parNames}{Vector or list of strings or epxressions of parameter names.
Must be list if expressions are used.}

\item{burn}{Integer specifying the num of iterations to discard of each
chain.}

\item{lags}{Integer specifying the num of lags upto which to compute the ACF.}
}
\description{
Plots the autocorrelation function (ACF) (see \code{\link[stats]{acf}}) per
MCMC chain. Plots are laid out in a grid with columns containing plots
grouped per MCMC method and rows containing the plots grouped by parameter.
}
