% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.R
\name{PlotHPDOverTime}
\alias{PlotHPDOverTime}
\title{Plot highest posterior density region over time}
\usage{
PlotHPDOverTime(statDraws, observations, dates, startDate = NULL,
  endDate = NULL, alpha = 0.99, ylab = NULL, statStr = "",
  obsStr = "Obs", statsML = NULL, zoomDate = NULL, xy1 = NULL,
  xy2 = NULL, modeCol = "red", MLCol = "blue", fillCol = "pink1",
  borderCol = "red2")
}
\arguments{
\item{statDraws}{Matrix of posterior samples of shape \code{numDraws} by
\code{numObs}, so each row represents one posterior sample of a time series
of the statistic.}

\item{observations}{Numeric vector of length \code{numObs} to plot as
observational proxy for the statistic.}

\item{dates}{Vector of strings of length \code{numObs} representing the
dates at which the statsitic is observed.}

\item{startDate}{String representing the starting date for the plot. In
combination with \code{endDate} it allows the user to specify a specific
date range contained in \code{dates} for display purposes. If NULL the
first date in dates is used. If not NULL it must be an element of dates.}

\item{endDate}{Same as \code{startDate}, but for the end of the date range.}

\item{alpha}{Numeric specifying the posterior probability mass contained in
the HPD region.}

\item{ylab}{Y axis label passed to \code{\link{plot}}.}

\item{statStr}{String to describe the statistic.}

\item{obsStr}{String to describe the observation series.}

\item{statsML}{Numeric vector of length \code{numObs} (Optional). Time series
of the statsitic's value under maximum likelihood (ML) parameters.}

\item{zoomDate}{String specifying a date for which to display a small
histogram of the posterior sample of the statistic in the top right corner
of the plot (Optional). If NULL, no such histogram is displayed.}

\item{xy1}{Location specifiers for drawing the top line from the
\code{zoomDate} inside the main plot to the histogram (Optional). If NULL,
no such line is displayed.}

\item{xy2}{Same as \code{xy1}, but for the bottom line.}

\item{modeCol}{Color for the mode. Passed to \code{col} argument for
\code{\link{plot}}.}

\item{MLCol}{Color for the \code{statsML}. Passed to \code{col} argument
for \code{\link{plot}}.}

\item{fillCol}{Color to fill the HPD region. Passed to \code{col} argument
for \code{\link{polygon}}.}

\item{borderCol}{Color for the borders of the HPD region. Passed to
\code{col} argument \code{\link{plot}}.}
}
\description{
Plots the (100 * \code{alpha})% highest posterior density (HPD) boundaries of
a statstic for which posterior samples are provided over a certain time
frame.
}
