% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.R
\name{PlotJointDists}
\alias{PlotJointDists}
\title{Scatter plots of joint posterior distributions}
\usage{
PlotJointDists(samples, paramPairs, lbParams, ubParams, burn = 1000)
}
\arguments{
\item{samples}{Matrix containing MCMC samples. Each column contains a
seperate chain of a parameter specified in \code{paramPairs} and is named
accordingly (e.g. given a \code{paramPair} c("A", "B") \code{samples} is
expected to contain a column named "A" and a column named "B" ).}

\item{paramPairs}{A list of \code{parNames} (see e.g.
\code{\link{PlotTraces}} for a description of a parNames object).}

\item{lbParams}{Named numeric vector representing the lower plot limits of
the parameters. Names correspond to the \code{parName}s specified in
\code{paramPairs}.}

\item{ubParams}{Named numeric vector representing the upper plot limits of
the parameters. Names correspond to the \code{parName}s specified in
\code{paramPairs}.}

\item{burn}{Integer specifying the num of iterations to discard of each
chain.}
}
\description{
Creates scatter plots of sample values of one parameter vs another.
}
