% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.R
\name{PlotTraces}
\alias{PlotTraces}
\title{Trace plots of MCMC samples}
\usage{
PlotTraces(samples, mcmcNames, parNames, iters = -(1:1000),
  cols = (1:8))
}
\arguments{
\item{samples}{Matrix containing MCMC samples. Each column contains a
seperate chain of a unique combination of a MCMC method specified in
\code{mcmcNames} and parameter specified in \code{parNames}. Columns
must be named according to the convention "\code{mcmcName} \code{parName}"
(e.g. "RWMH A").}

\item{mcmcNames}{Vector or list of strings or epxressions of MCMC method
names. Must be list if expressions are used.}

\item{parNames}{Vector or list of strings or epxressions of parameter names.
Must be list if expressions are used.}

\item{iters}{Indexing vector used to select the iterations to plot.}

\item{cols}{Vector of elements that can be passed to \code{\link{plot}} as
\code{col} (color) values applied per parameter. Elements are repeated to
match length of parNames.}
}
\description{
Creates plots of iterations vs. sampled values for each chain. Plots are laid
out in a grid with columns containing plots grouped per MCMC method and rows
containing the plots grouped by parameter.
}
