% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_gas.R
\name{RWMH}
\alias{RWMH}
\title{Random Walk Metropolis Hastings.}
\description{
Function of similar form as \code{GGS}, but draws are generated using a
Random Walk Metropolis Hasting's algorithm with multivariate Student's-t
proposal density.
}
