% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_gas.R
\name{PriorStack}
\alias{PriorStack}
\alias{Rcpp_PriorStack-class}
\alias{Rcpp_PriorStack}
\title{An object that fully specifies the prior for a model.}
\description{
Allows for flexible specification of the prior over all model parameters.
Currently supports Normal, TruncatedNormal & ImproperUniform.
}
\section{Usage}{

\preformatted{
p <- new(PriorStack, priorStrs, priorParams)
p <- new(PriorStack, priorStrs, priorParams, priorToParamIndex)

p$LogPriors(params)
p$LogPriorsWPar(params, priorParams)
p$GradLogPriors(params)

p$Filter(y, f1)
}
}

\section{Details}{

}

