% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_gas.R
\name{RWMH}
\alias{RWMH}
\title{Random Walk Metropolis Hastings.}
\usage{
\preformatted{
RWMH(modelStr, priorStack, y, f1, initParams, sigma, iter, stepsize, df,
  verbose, printIter, thinning)
}
}
\description{
Function of similar form as \code{\link{GGS}}, but draws are generated using
a Random Walk Metropolis Hasting's algorithm with multivariate Student's-t
proposal density.
}
