% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_gas.R
\name{VectorizedPosterior}
\alias{VectorizedPosterior}
\title{An object that fully specifies the prior for a model.}
\description{
Allows for flexible specification of the prior over all model parameters.
}
